requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
opencv-python==4.7.0.76
matplotlib==3.6.3
numpy==1.24.1
python-docx==0.8.11
fastapi[all]==0.95.2